refactor: use shared component to select fields from header
