+++ fdm-app/app/components/blocks/header/balance.tsx
@@ -8,6 +8,7 @@ import {
   DropdownMenuContent,
   DropdownMenuTrigger,
 } from "~/components/ui/dropdown-menu"
+import { HeaderFieldPicker } from "~/components/blocks/header/field-picker"
 
 export function HeaderBalance({
   b_id_farm,
@@ -21,6 +22,7 @@ export function HeaderBalance({
   const location = useLocation()
   const currentPath = String(location.pathname)
   const calendar = useCalendarStore((state) => state.calendar)
+  const balanceKind = currentPath.includes("/balance/nitrogen") ? "nitrogen" : "organic-matter"
 
   return (
     <>
@@ -59,28 +61,11 @@ export function HeaderBalance({
         <>
           <BreadcrumbSeparator />
           <BreadcrumbItem>
-            <DropdownMenu>
-              <DropdownMenuTrigger className="flex max-w-[120px] items-center gap-1 outline-none sm:max-w-[200px] md:max-w-none">
-                <span className="truncate">
-                  {fieldOptions
-                    ? (fieldOptions.find((option) => option.b_id === b_id)?.b_name ??
-                      "Unknown field")
-                    : "Kies een perceel"}
-                </span>
-                <ChevronDown className="text-muted-foreground h-4 w-4 shrink-0" />
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="start">
-                {fieldOptions.map((option) => (
-                  <DropdownMenuCheckboxItem checked={b_id === option.b_id} key={option.b_id}>
-                    <NavLink
-                      to={`/farm/${b_id_farm}/${calendar}/balance/${currentPath.includes("nitrogen") ? "nitrogen" : "organic-matter"}/${option.b_id}`}
-                    >
-                      {option.b_name}
-                    </NavLink>
-                  </DropdownMenuCheckboxItem>
-                ))}
-              </DropdownMenuContent>
-            </DropdownMenu>
+            <HeaderFieldPicker
+              b_id={b_id}
+              fieldOptions={fieldOptions}
+              buildHref={(optionId) => `/farm/${b_id_farm}/${calendar}/balance/${balanceKind}/${optionId}`}
+            />
           </BreadcrumbItem>
         </>
       ) : null}
