+++ fdm-app/app/components/blocks/header/indicators.tsx
@@ -1,5 +1,4 @@
-import { ChevronDown } from "lucide-react"
-import { NavLink, useMatches, useParams } from "react-router"
+import { useMatches, useParams } from "react-router"
 import { useCalendarStore } from "@/app/store/calendar"
 import {
   BreadcrumbItem,
@@ -7,12 +6,7 @@ import {
   BreadcrumbPage,
   BreadcrumbSeparator,
 } from "~/components/ui/breadcrumb"
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "~/components/ui/dropdown-menu"
+import { HeaderFieldPicker } from "~/components/blocks/header/field-picker"
 
 export function HeaderIndicators({ b_id_farm }: { b_id_farm: string }) {
   const calendarFromStore = useCalendarStore((state) => state.calendar)
@@ -47,23 +41,11 @@ export function HeaderIndicators({ b_id_farm }: { b_id_farm: string }) {
           <BreadcrumbSeparator />
           <BreadcrumbItem>
             {fieldList.length > 1 ? (
-              <DropdownMenu>
-                <DropdownMenuTrigger className="focus-visible:ring-ring flex max-w-30 items-center gap-1 outline-none focus-visible:ring-2 focus-visible:ring-offset-2 sm:max-w-50 md:max-w-none">
-                  <span className="truncate">{fieldName ?? b_id}</span>
-                  <ChevronDown className="h-4 w-4 shrink-0" />
-                </DropdownMenuTrigger>
-                <DropdownMenuContent align="start" className="max-h-72 overflow-y-auto">
-                  {fieldList.map((f) => (
-                    <DropdownMenuItem
-                      key={f.b_id}
-                      asChild
-                      className={f.b_id === b_id ? "font-semibold" : ""}
-                    >
-                      <NavLink to={`${basePath}/${f.b_id}`}>{f.b_name ?? f.b_id}</NavLink>
-                    </DropdownMenuItem>
-                  ))}
-                </DropdownMenuContent>
-              </DropdownMenu>
+              <HeaderFieldPicker
+                b_id={b_id}
+                fieldOptions={fieldList}
+                buildHref={(optionId) => `${basePath}/${optionId}`}
+              />
             ) : (
               <BreadcrumbPage className="max-w-30 truncate sm:max-w-50 md:max-w-none">
                 {fieldName ?? b_id}
