+++ fdm-app/app/components/blocks/header/nutrient-advice.tsx
@@ -1,13 +1,6 @@
-import { ChevronDown } from "lucide-react"
-import { NavLink } from "react-router"
 import { useCalendarStore } from "@/app/store/calendar"
 import { BreadcrumbItem, BreadcrumbLink, BreadcrumbSeparator } from "~/components/ui/breadcrumb"
-import {
-  DropdownMenu,
-  DropdownMenuCheckboxItem,
-  DropdownMenuContent,
-  DropdownMenuTrigger,
-} from "~/components/ui/dropdown-menu"
+import { HeaderFieldPicker } from "~/components/blocks/header/field-picker"
 
 export function HeaderNutrientAdvice({
   b_id_farm,
@@ -32,26 +25,11 @@ export function HeaderNutrientAdvice({
         <>
           <BreadcrumbSeparator />
           <BreadcrumbItem>
-            <DropdownMenu>
-              <DropdownMenuTrigger className="focus-visible:ring-ring flex max-w-30 items-center gap-1 outline-none focus-visible:ring-2 focus-visible:ring-offset-2 sm:max-w-50 md:max-w-none">
-                <span className="truncate">
-                  {fieldOptions
-                    ? (fieldOptions.find((option) => option.b_id === b_id)?.b_name ??
-                      "Unknown field")
-                    : "Kies een perceel"}
-                </span>
-                <ChevronDown className="text-muted-foreground h-4 w-4 shrink-0" />
-              </DropdownMenuTrigger>
-              <DropdownMenuContent align="start">
-                {fieldOptions.map((option) => (
-                  <DropdownMenuCheckboxItem checked={b_id === option.b_id} key={option.b_id}>
-                    <NavLink to={`/farm/${b_id_farm}/${calendar}/nutrient_advice/${option.b_id}`}>
-                      {option.b_name}
-                    </NavLink>
-                  </DropdownMenuCheckboxItem>
-                ))}
-              </DropdownMenuContent>
-            </DropdownMenu>
+            <HeaderFieldPicker
+              b_id={b_id}
+              fieldOptions={fieldOptions}
+              buildHref={(optionId) => `/farm/${b_id_farm}/${calendar}/nutrient_advice/${optionId}`}
+            />
           </BreadcrumbItem>
         </>
       ) : null}
